requirements.txt
ffmpeg-python
scipy
matplotlib